api/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
honcho==1.1.0
pytz==2022.7.1
sqlparse==0.4.3
